backend folder is ready for docker
Type in the command 'docker-compose up' in the backend folder CLI. The database with 5 income tax bracket documents will be initialized. Any changes to the database can be done through API calls.

THIS VERSION CANNOT BE USED FOR DEVELOPMENT. TO DO SO, GO TO APPLICATION.PROPERTIES AND CHANGE mongodb AT THE LAST LINE TO localhost